requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.0.3
numpy==1.24.3
Werkzeug==2.3.7
supabase==2.0.0
python-dotenv==1.0.0
prometheus-flask-exporter==0.23.0
redis==5.0.0
psutil==5.9.5
plotly==5.17.0